Create the new module namespace for the server.
I decided to create a new namespace for the application for the modules and
handlers.  Yes, it's a bit weird, but that's because of the name.  Anyway,
create the namespace, and moved three of the former modules (non-handlers)
to the new namespace, and update the code to point to them.