Update and rename README to README.md
A nice and shiney new README for the Github main repo page :) Not intended to actually replace the text version for distribution, but should look really nice. I hope you like it :) Maybe I can do more docs later on operations and collab that w/you for a pseudo-sort of manual.