Refactor CGI and SCGI code.
I moved the environment creation to its own function so ensure consistency
between CGI and SCGI environments.

I also got rid of the cwd directive.  It was kind of a nice idea but I
decided against it because ...  I changed my mind for reasons I just can't
explain.  So be it.