Bug fix---filesystem traversal of incorrect request
As we descend the filesystem, and we come across the file, check to see if
there's more to the request path---there shouldn't be for a file (there can
be for a CGI/SCGI).  If it's not a CGI/SCGI and there is more to the request
path, signal an error for buggy clients.