Paths are now strings, not patterns.
This simplifies the configuration as you no longer have to specify a pattern
to match against, just a leading string.  Also, the handlers no longer have
to be specified in order as that is taken care of during initialization,
where they are sorted longest match wins.  Also, the filesystem handler will
now do a permanent redirect if the URL ends in a directory without a
trailing slash (no longer have to specify that in the permanent redirect
section).

Also, charset and language have a default value, and are cascaded down
through the configuration.